// @types/types.d.ts
// Generated by https://quicktype.io

export interface IPage {
  object: string;
  id: string;
  created_time: string;
  last_edited_time: string;
  created_by: CreatedBy;
  last_edited_by: CreatedBy;
  has_children: boolean;
  archived: boolean;
  type: string;
  child_page: ChildPage;
}

export interface ChildPage {
  title: string;
}

export interface CreatedBy {
  object: string;
  id: string;
}

export interface IDatabase {
  object: string;
  id: string;
  created_time: string;
  last_edited_time: string;
  created_by: CreatedBy;
  last_edited_by: CreatedBy;
  cover: any;
  icon: IIcon;
  parent: Parent;
  archived: boolean;
  properties: IBaseProperties;
  url: string;
}

export interface Parent {
  type: string;
  database_id: string;
}

export interface IBaseProperties {
  url: Iurl;
  Name: IName;
  Type: IType;
  Description: IDescription;
  icon: IIcon;
}

interface IName {
  title: {
    text: {
      content: any;
    };
  }[];
}

export type ISelectOp = {
  id: string;
  name: string;
  color: string;
};

interface IType {
  multi_select: ISelectOp[];
}

interface Iurl {
  rich_text: IRichText;
}

interface IDescription {
  rich_text: IRichText;
}

type IRichText = { plain_text: string }[];

type IIcon = IEmoji;

interface IEmoji {
  type: 'emoji';
  emoji: string;
}